feat: calculate the commission

+++ src/main/java/com/bayzdelivery/model/Delivery.java
@@ -2,6 +2,8 @@ package com.bayzdelivery.model;
 
 import java.io.Serializable;
 import java.time.Instant;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -37,7 +39,7 @@ public class Delivery implements Serializable{
   Long price;
 
   @Column(name = "comission")
-  Long comission;
+  Double comission;
 
   @ManyToOne
   @JoinColumn(name = "delivery_man_id", referencedColumnName = "id")
@@ -87,11 +89,11 @@ public class Delivery implements Serializable{
     this.price = price;
   }
 
-  public Long getComission() {
+  public Double getComission() {
     return comission;
   }
 
-  public void setComission(Long comission) {
+  public void setComission(Double comission) {
     this.comission = comission;
   }
